requirements.txt
pandas==2.0.3
numpy==1.24.3
requests==2.31.0
aiohttp==3.8.5
beautifulsoup4==4.12.2
openpyxl==3.1.2
fastapi==0.103.0
uvicorn==0.23.2
python-dotenv==1.0.0
selenium==4.12.0
webdriver-manager==4.0.0
